requirements.txt
pyautogen
llama-index
llama-index-tools-wikipedia
llama-index-readers-wikipedia
wikipedia
streamlit